backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pandas==2.2.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
